# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Тропа памяти: Последний бой — API",
    "version": "0.1.0",
    "description": "Контракт, который ожидает фронтенд. Источник правды — frontend/src/contract (zod). Сгенерировано: npm run contract."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Тот же домен, что и фронт (прокси)"
    }
  ],
  "paths": {
...
    "/routes": {
      "get": {
        "operationId": "listRoutes",
        "summary": "Семейные маршруты",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Route"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      }
    },
    "/routes/{id}": {
      "get": {
        "operationId": "getRoute",
        "summary": "Маршрут с точками",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/family/fighters": {
      "get": {
        "operationId": "listFamilyFighters",
        "summary": "Бойцы семьи вошедшего пользователя",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyFighter"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      },
      "post": {
        "operationId": "createFamilyFighter",
        "summary": "Добавить бойца в семейный архив",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyFighter"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyFighterInput"
              }
            }
          }
        }
      }
    },
    "/family/fighters/{id}": {
      "patch": {
        "operationId": "updateFamilyFighter",
        "summary": "Исправить сведения о бойце (записи не меняются)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyFighter"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyFighterInput"
              }
            }
          }
        }
      }
    },
    "/family/fighters/{id}/delete": {
      "post": {
        "operationId": "deleteFamilyFighter",
        "summary": "Удалить бойца вместе с найденными записями",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/family/fighters/{id}/records": {
      "post": {
        "operationId": "addFamilyRecord",
        "summary": "Добавить найденную запись: только ссылки на официальные базы, без повторов",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyFighter"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoundRecordInput"
              }
            }
          }
        }
      }
    },
    "/family/fighters/{id}/records/{recordId}/delete": {
      "post": {
        "operationId": "deleteFamilyRecord",
        "summary": "Удалить найденную запись",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyFighter"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "LatLon": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "additionalProperties": false,
        "description": "Координаты WGS84"
      },
      "SourceKind": {
        "type": "string",
        "enum": [
          "book_of_memory",
          "obd_memorial",
          "pamyat_naroda",
          "osm",
          "archive",
          "literature",
          "eyewitness",
          "demo"
        ],
        "description": "Тип источника факта"
      },
      "Source": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "kind",
          "title"
        ],
        "additionalProperties": false,
        "description": "Источник факта. У каждого факта в интерфейсе есть хотя бы один источник"
      },
...
      "PointKind": {
        "type": "string",
        "enum": [
          "battle",
          "trench",
          "hq"
        ],
        "description": "Тип точки маршрута: battle — Звезда (бой), trench — Шлем (окоп), hq — Книга (штаб)"
      },
      "ChildTask": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1
          },
          "options": {
            "minItems": 2,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "answerIndex": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "illustration": {
            "description": "Путь к иллюстрации в ассетах фронта",
            "type": "string"
          },
          "explanation": {
            "type": "string",
            "minLength": 1,
            "description": "Что сказать после ответа"
          }
        },
        "required": [
          "question",
          "options",
          "answerIndex",
          "explanation"
        ],
        "additionalProperties": false,
        "description": "Задание для ребёнка на точке маршрута"
      },
      "RoutePoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "kind": {
            "$ref": "#/components/schemas/PointKind"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "story": {
            "type": "string",
            "minLength": 1
          },
          "task": {
            "$ref": "#/components/schemas/ChildTask"
          },
          "sources": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Source"
            }
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "lat",
          "lon",
          "story",
          "task",
          "sources"
        ],
        "additionalProperties": false,
        "description": "Интерактивная точка маршрута"
      },
      "Route": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "summary": {
            "type": "string"
          },
          "lengthM": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "durationMin": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "path": {
            "minItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LatLon"
            }
          },
          "points": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            }
          },
          "demo": {
            "type": "boolean",
            "description": "true — демо-данные, вымышлены или не проверены"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "lengthM",
          "durationMin",
          "path",
          "points",
          "demo"
        ],
        "additionalProperties": false,
        "description": "Семейный квест-маршрут. Точки упорядочены, path — линия маршрута"
      },
...
      "FoundRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "id",
          "url",
          "title"
        ],
        "additionalProperties": false,
        "description": "Найденный документ — ссылка на страницу в «Памяти народа», ОБД «Мемориал» или «Подвиге народа»"
      },
      "FamilyFighter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "firstName": {
            "type": "string",
            "maxLength": 60
          },
          "middleName": {
            "type": "string",
            "maxLength": 60
          },
          "birthYear": {
            "type": "integer",
            "minimum": 1860,
            "maximum": 1935
          },
          "relation": {
            "type": "string",
            "maxLength": 60,
            "description": "Кем приходится: «прадед по маме»"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "description": "Что известно в семье"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoundRecord"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "Дата и время ISO 8601, UTC"
          }
        },
        "required": [
          "id",
          "lastName",
          "firstName",
          "middleName",
          "relation",
          "note",
          "records",
          "createdAt"
        ],
        "additionalProperties": false,
        "description": "Боец семьи в личном архиве пользователя. Виден только владельцу (нужен вход)"
      },
      "FamilyFighterInput": {
        "type": "object",
        "properties": {
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "firstName": {
            "type": "string",
            "maxLength": 60
          },
          "middleName": {
            "type": "string",
            "maxLength": 60
          },
          "birthYear": {
            "type": "integer",
            "minimum": 1860,
            "maximum": 1935
          },
          "relation": {
            "type": "string",
            "maxLength": 60,
            "description": "Кем приходится: «прадед по маме»"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "description": "Что известно в семье"
          }
        },
        "required": [
          "lastName",
          "firstName",
          "middleName",
          "relation",
          "note"
        ],
        "additionalProperties": false,
        "description": "Новый боец или правка: сервер проверяет те же правила, что и форма (domain/familyArchive)"
      },
      "FoundRecordInput": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "title": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "url",
          "title"
        ],
        "additionalProperties": false,
        "description": "Ссылка на найденный документ; без названия — название базы"
      },
...
      "Ack": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false,
        "description": "Подтверждение действия"
      },
...
    }
  }
}